Updated save and load user data and added photo thumbnails

+++ src/photosfx/controller/AlbumController.java
@@ -1,7 +1,7 @@
 package photosfx.controller;
 
 import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.io.*;
 
 import javafx.event.ActionEvent;
@@ -10,8 +10,9 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.*;
-import javafx.scene.layout.VBox;
+import javafx.scene.layout.*;
 import javafx.stage.*;
+import javafx.scene.image.*;
 import photosfx.model.*;
 
 public class AlbumController {
@@ -35,40 +36,107 @@ public class AlbumController {
     private Button recaptionButton;
 
     @FXML
+    private Button addTagsButton;
+
+    @FXML
     private VBox photosContainer;
+
+    @FXML
+    private Button logoutButton;
+
+    @FXML
+    private Button homeButton;
+
+    @FXML
+    private Label captionLabel;
+
+    @FXML
+    private Label dateTimeLabel;
+
+    @FXML
+    private Label tagsList;
+
+    @FXML
+    private FlowPane thumbnailsFlowPane;
+
+    @FXML
+    private VBox photoDisplayVBox;
+
+    @FXML
+    private ImageView photoDisplayImageView;
 
     private Album album;
 
-    public void initialize(Album album) {
+    private User user;
+
+    public void initialize(Album album, User user) {
         this.album = album;
+        this.user = user;
         if (album == null) {
             System.out.println("album load failed");
         }
+        if (user == null) {
+            System.out.println("user load failed");
+            return;
+        }
         System.out.println(album.getAlbumName() + "loaded");
+        System.out.println("User is " + user.getUsername());
         showPhotos();
     }
 
     private void showPhotos() {
+        System.out.println("User is " + this.user.getUsername() + " and album is " + this.album.getAlbumName());
+        String userPath = DataFileManager.basePath + File.separator + user.getUsername();
+        String albumPath = userPath + File.separator + album.getAlbumName();
+        user.updateAlbum(album);
+        System.out.println("Photos up to date...");
         photosContainer.getChildren().clear();
         List<Photo> photos = album.getPhotos();
-
         if (photos == null) {
             System.out.println("Photos is null");
             return;
         }
         for (Photo photo : photos) {
+            File photoFile = new File(photo.getPathInDisk());
+            if (!photoFile.exists()) {
+                System.out.println("Photo file does not exist: " + photo.getPathInDisk());
+                continue;
+            }
+            Image thumbnailImage = new Image(photoFile.toURI().toString());
+            ImageView thumbnailView = new ImageView(thumbnailImage);
+            thumbnailView.setFitWidth(100); // Set thumbnail width
+            thumbnailView.setPreserveRatio(true); // Preserve aspect ratio
+            thumbnailView.setSmooth(true); // Enable smoother image rendering
+            thumbnailView.setCache(true); // Cache the image for better performance
+
+            thumbnailView.setOnMouseClicked(event -> displayFullSizeImage(photo));
+
             Button deleteButton = new Button("Delete");
             deleteButton.setOnAction(event -> deletePhoto(photo));
 
-            Button renameButton = new Button("Rename");
+            Button recaptionButton = new Button("Rename");
             recaptionButton.setOnAction(event -> recaptionPhoto(photo));
+
+            Button addTagsButton = new Button("Add tag");
+            addTagsButton.setOnAction(event -> addTag(photo));
 
             Label photoLabel = new Label(photo.getCaption());
             Label tagsLabel = new Label(photo.getTags().toString());
+            Label dateLabel = new Label(photo.getDate());
 
-            VBox photoBox = new VBox(photoLabel, tagsLabel, deleteButton, renameButton);
+            VBox photoBox = new VBox(thumbnailView, photoLabel, tagsLabel, dateLabel, deleteButton, recaptionButton,
+                    addTagsButton);
+            photoBox.setSpacing(5);
             photosContainer.getChildren().add(photoBox);
         }
+    }
+
+    private void displayFullSizeImage(Photo photo) {
+        Image fullImage = new Image(photo.getPathInDisk());
+        photoDisplayImageView.setImage(fullImage);
+        captionLabel.setText(photo.getCaption());
+        dateTimeLabel.setText(photo.getDate().toString());
+        tagsList.setText(photo.getTags().toString());
     }
 
     private void deletePhoto(Photo photo) {
@@ -80,7 +148,6 @@ public class AlbumController {
             showAlert("Error", "Failed to delete album.");
         }
         showPhotos();
-        ;
     }
 
     private void recaptionPhoto(Photo photo) {
@@ -99,6 +166,26 @@ public class AlbumController {
                 showAlert("Error", "Please enter a valid caption.");
             }
         });
+        showPhotos();
+    }
+
+    private void addTag(Photo photo) {
+        TextInputDialog dialog = new TextInputDialog(photo.getTags().toString());
+        dialog.setTitle("Add tags");
+        dialog.setHeaderText(null);
+        dialog.setContentText("Enter tag type and value. Eg: Location, New Brunswick");
+
+        dialog.showAndWait().ifPresent(newTag -> {
+            if (!newTag.isEmpty()) {
+                String[] tag = newTag.split(" ");
+                photo.setTag(tag[0], tag[1]);
+                showPhotos(); // Refresh album list
+                showAlert("Success", "Tag added.");
+            } else {
+                showAlert("Error", "Please enter a valid Tag.");
+            }
+        });
+        showPhotos();
     }
 
     private void showAlert(String title, String message) {
@@ -123,17 +210,62 @@ public class AlbumController {
             Optional<String> result = dialog.showAndWait();
             result.ifPresent(caption -> {
                 // Create a new Photo instance with the selected file and entered caption
-                Photo photo = new Photo(selectedFile.getName(), null); // You should adjust this to get the photo date
+                Calendar currentDate = Calendar.getInstance();
+                Photo photo = new Photo(selectedFile.getName(), currentDate); // You should adjust this to get the photo
+                                                                              // date
                 photo.setCaption(caption);
+                photo.setPath(selectedFile.getAbsolutePath());
 
                 // Add the photo to the album
                 album.addPhoto(photo);
+                System.out.println("album name for saving photo: " + album.getAlbumName());
+                String userPath = DataFileManager.basePath + File.separator + user.getUsername();
+                String albumPath = userPath + File.separator + album.getAlbumName();
+                DataFileManager.saveUser(this.user);
 
                 // Refresh the photos display
                 showPhotos();
 
                 showAlert("Success", "Photo added successfully.");
             });
+        }
+    }
+
+    @FXML
+    private void homeButtonClicked() {
+        try {
+            FXMLLoader loader = new FXMLLoader(getClass().getResource("../view/UserHome.fxml"));
+            Parent root = loader.load();
+            UserHomeController controller = loader.getController();
+            if (user == null) {
+                System.out.println("User not loaded");
+            }
+            controller.initialize(user); // Pass username to UserHomeController
+            Stage stage = new Stage();
+            stage.setScene(new Scene(root));
+            stage.setTitle("User Home");
+            stage.show();
+
+            // Close the current login window
+            Stage currentStage = (Stage) addPhotoButton.getScene().getWindow();
+            currentStage.close();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    @FXML
+    private void logoutButtonClicked() {
+        try {
+            FXMLLoader loader = new FXMLLoader(getClass().getResource("../view/Login.fxml"));
+            Parent root = loader.load();
+            Scene scene = new Scene(root);
+            Stage stage = (Stage) logoutButton.getScene().getWindow(); // Assuming logoutButton is the button that
+                                                                       // triggered the action
+            stage.setScene(scene);
+            stage.show();
+        } catch (IOException e) {
+            e.printStackTrace(); // Handle error appropriately
         }
     }
